refactor: move settings cards to subdir
