+++ app/settings/cards/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as ThemeSettingCard } from "./ThemeSettingCard";
